+++ sbi/neural_nets/estimators/flowmatching_estimator.py
@@ -9,7 +9,7 @@ import torch.nn as nn
 from torch import Tensor
 
 from sbi.neural_nets.estimators.base import ConditionalVectorFieldEstimator
-from sbi.neural_nets.net_builders.vector_field_nets import VectorFieldNet
+from sbi.utils.vector_field_utils import VectorFieldNet
 
 
 class FlowMatchingEstimator(ConditionalVectorFieldEstimator):
