+++ sbi/neural_nets/net_builders/vector_field_nets.py
@@ -22,27 +22,7 @@ from sbi.utils.sbiutils import (
     z_standardization,
 )
 from sbi.utils.user_input_checks import check_data_device
-
-
-class VectorFieldNet(nn.Module, ABC):
-    """Abstract base class for vector field estimation networks.
-
-    Used by both flow matching and score matching approaches.
-    """
-
-    @abstractmethod
-    def forward(self, theta: Tensor, x: Tensor, t: Tensor) -> Tensor:
-        """Forward pass to compute the vector field.
-
-        Args:
-            theta: Parameters
-            x: Conditioning information
-            t: Time parameter (scalar or batched)
-
-        Returns:
-            Vector field evaluation at the provided points
-        """
-        pass
+from sbi.utils.vector_field_utils import VectorFieldNet
 
 
 # ==================== Building Flow/Score Matching Estimators =========================
